+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 16.10.11
 - QOwnNotes is now built with Qt 5.7 under OS X and Windows instead of Qt 5.5
+- added information about the system environment to the debug settings dump
 
 ## 16.10.10
 - the controls in the log panel are now hidden by default and can be